refactoring: components renamed and app closing process updated to include databases closing.
The ‘Article’ component renamed to ‘Links’ and the ‘SqliteUrls’ renamed to ‘Bookmarks’. Their stylesheets and templates renamed accordingly.  The ‘SiteMenuItems’ array in the ‘angObjects.ts’ and the ‘app.route.ts’ also updated correspondingly.

The app closing process updated to include the graceful closing of the working and main databases. The function in charge is the ‘closeDBConnections’ (its previous name was ‘closeMainConnection’) function in the electron/dbs/sqlite/connections.ts file. The 'sqlite:close-db-connections' IPC channel handler added into the electron/main.ts file. The app closing process is triggered by the ‘onExitClick()’ function in the Layout component (via clicking on the header exit icon), which calls the ‘closeDbConnections()’ function in the BackEnd service