refact: A new scrapeMediumMarkdownContent() function was added and integrated into collectPostsFromUrlTabs(), enabling full Markdown content extraction alongside metadata. All time-related delays were moved to a shared time-constants.ts file for consistency.
All hardcoded time-related values (e.g. delays) used in functions inside the ‘scrape-functions.ts’ and ‘page-convert.ts’ files, are substituted by corresponding time-related constants, that are defined into the new file: ‘time-connstants.ts’.

The new function ‘scrapeMediumMarkdownContent()’ added into the ‘scrape-functions.ts’ file. This function is based on the ‘htmlToMarkdown()’ function in the ‘page-converter.ts’ file, and it also uses the ‘turndown’ package (TurndownService), and also calls the helper functions: ‘getCleanedPageContent()’, ‘processGists()’, and ‘extractCodeFromIframe()’.

The new ‘scrapeMediumMarkdownContent()’ function is called by the existing main ‘wrapper’ function: ‘collectPostsFromUrlTabs()’, in conjunction with the also existing ‘scrapeMediumArticle()’ function. Now the ‘collectPostsFromUrlTabs()’ functions returns a PostData array including the (markdown) content in each PostData object.

The Cloudflare challenge detection is also added in the beginning of the key ‘scrapeMediumArticle()’ function.

The HtmlMarkdown Component, has been updated: The ‘srapeBasicData()’ function renamed to ‘srapeArticleData()’, and it is the only function being used to scrape article data including the markdown content. (The'convert-html-to-markdown' IPC invoker is not used anymore, so, the corresponding htmlToMarkdown function is also not used anymore )

Documentation files: ‘scrape-functions-full-doc.md’ and ‘page-converters-full-doc.md’ have been added/updated