feat/refact: Added URL existence check in main DB with new IPC handler, updated Angular BackEnd service to expose DB queries, enhanced dialog auto-close, and refactored Markdown component to validate URLs and insert articles via the service.
feat:
The new function: ‘isUrlExisting()’ that checks if a URL already exists in the 'articles' table, has been added in the maindb_queries.ts file. The corresponding IPC handler 'sqlite:check-if-url-existing' method has been also added into the Electron main.ts file.

refact:
The Angular BackEnd Service is refactored and updated to additionally provide access to Electron main DB queries. The ‘insertArticles()’ and ‘checkUrlExists()’ functions have added to the BackEnd Service. These 2 functions use the service generic invoker function ‘ipcInvoke()’ for invoking the ‘sqlite:insert-articles-from-json-array’ and ‘sqlite:check-if-url-existing’ IPC handlers defined into the main.ts.

The Dialog component has been updated to also include the ‘warn’ token for auto closing the corresponding pop-up window after the value of ms passed in with the delay property.

The Markdown Component has been updated/refactored. The new function ‘isUrlExisting()’ has been added for checking if a URL entered in the form field ‘url’ does exists or not in the articles table of the main DB. It does it, by using the ‘checkUrlExists()’ BackEnd Service function. The newly added ‘isUrlExisting()’ function is called from within the ngOnInit() component lifecycle hook method (Actually from within the subscription that checks the url changes). The ‘isUrlExisting()’ function returns a Promise<boolean> and it is called as Observable using the from(this.isUrlExisting(this.linkURL())).subscribe({… . The ‘from(Promise)…’ completes automatically, so no manual unsubscribe needed. The url is cleaned (the query parameters part is removed – after the ? question mark) before it is assigned as value to the ‘linkURL()’ signal and passed as a parameter into the ‘isUrlExisting()’ function. In case that a cleaned url exists then a ‘warn’ pop-up window is displayed via the instantiation of the Dialog Component, which auto-close after .5 sec.

The Markdown Component function insertScrapedArrayToDB() has been also updated and now does not call directly the 'sqlite:insert-articles-from-json-array' IPC channel handler. It calls it indirectly via calling the ‘insertArticles()’ method of the BackEnd Service.