fix/refact: Check the Medium Slug existence instead of a full URL, to handle cases where a ‘Cleaned’ URL is changed after its page is opened on the host/site.
The new function ‘getMediumSlugFromUrl()’ was added to the ‘shared-utils.ts’ file. It extracts the Medium Slug for a given URL.

The new function ‘isSlugExisting()’ was added to ‘maindb_queries.ts’,  to search if the linkurl column contains the provided string, not just an exact full URL match. It now uses LIKE with wildcards, instead of = for the search. The main.ts updated with the new IPC channel handler: 'sqlite:check-if-slug-exists' that calls the ‘isSlugExisting()’ function. The BackEnd Service is also updated with the new ‘checkSlugExists()’ function that deals with the 'sqlite:check-if-slug-exists' channel.

The Markdown Component has been updated. The function ‘isUrlExisting()’ renamed to ‘isSlugExisting()’. First, it extracts the Medium slug from the URL using the ‘getMediumSlugFromUrl()’ function, and then calls the  BackEnd Service function ‘checkSlugExists()’ passing it the extracted slug as a parameter. In the ngOnInit, the from block: ‘from(this.isSlugExisting(this.linkURL())).subscribe({…’ updated accordingly.