feat: If an article URL slug already exists (is already inserted) into main DB, then fetch and show it, in Markdown Component. Also, indicate if an Article is newly inserted, or it already exists
The new function ‘getPostBySlug()’ added in the ‘electron/dbs/sqlite/maindb_queries.ts’ file. It returns a PostData object or null.

The new IPC channel handler 'sqlite:get-post-data-by-slug', added in the Electron main.ts.

The new ‘getPostDatabySlug()’ added in the Angular BackEnd Service It calls the IPC channel handler 'sqlite:get-post-data-by-slug' via the generic ipcInvoke<T>() BackEnd service private method.

In the Markdown Component, the new function ‘getArticleDataBySlug()’ added. It calls the function ‘getPostDatabySlug()’ in the BackEnd Service. The new function ‘getArticleDataBySlug()’ is called as an Observable using the ‘from(Promise)’.

The new method ‘showArticleData()’ also added in the Markdown Component. It sets the ‘postMetaDataString()’ and ‘markdownString()’ signals and triggers the text to be shown in the corresponding text areas. It is also called from the ‘srapeArticleData()’ function.

Finally, a new span is added in the Markdown template to indicate, using some appropriate text, if the content concerns a new article or an existing one. Corresponding settings added into the Narkdown stylesheet file.