feat: the NG-Zorro Tree and NzTreeModule used for providing categories / subcategories array of nodes within the ‘sider-menu’ middle area of the sider in the nz-layout of the DbArticles Component.
‘varObjects.ts’
The new interfaces ‘Category’ and ‘CategoryNode’, as well as the type: ‘CategoryRow’ added.

‘maindb_queries.ts’
The new functions ‘getSubcategoryTree()’ and ‘getSubcategoryForest()’ added. The ‘getSubcategoryForest()’ function builds and returns a forest (: an array of root CategoryNode) for all categories for a given `parent_Id`. Especially, if the parent_Id === null, then, the forest for the top-level categories (roots) is returned.

‘electron-utils.ts’
The ‘coerceParameter()’ function added for coercing the type of an input parameter passed in. It is used in the new functions added into the ‘maindb_queries.ts’ file.

‘main.ts’
The IPC channel handlers 'sqlite:get-categories-by-parent-id' and 'sqlite:get-sub-category-forest-by-parent-id' added, for the ‘getSubcategoryTree()’ and ‘getSubcategoryForest()’ functions respectively.

‘BackEnd’ Service (‘back-end.ts’ file)
The functions: ‘getCategoriesByParentId()’ and  ‘getCategoryForestByParentId()’ added for calling the IPC channel handlers 'sqlite:get-categories-by-parent-id' and 'sqlite:get-sub-category-forest-by-parent-id' respectively.

‘DbArticles’ Component.
The NG-Zorro Tree and NzTreeModule used for providing categories / subcategories array of nodes within the ‘sider-menu’ middle area of the sider in the nz-layout of the DbArticles Component. The async functions ‘getCategoriesByParentId()’ and ‘getCategoryForestByParentId()’ added that call the BackEnd Service methods: ‘getCategoriesByParentId()’ and  ‘getCategoryForestByParentId()’ respectively.
The function ‘mapCategoryNodesToTree()’ added and it is used to map the forest of the categories / subcategories array fetched by the ‘getCategoryForestByParentId()’ function to an ‘NzTreeNodeOptions’ array of keys and titles to be used as the [nzData] with the <nz-tree> tag.
The ‘Root Only’ nz-checkbox in the sider-footer, renamed to ‘Unassigned’. Checking/Unchecking it cause the nz-tree to be disabled or enabled respectively.
Various updates and tweaks made into the stylesheet for the output UI and its behavior.