feat/debug: initial implementation of a Table (‘NzTableModule’) to present data from backend
‘DbArticles’ Component (‘db-articles.ts’ file).
It is updated to use the ‘NzTableModule’.  The Articles’ meta data fetched from backend (via ‘getUnassignedArticles()’) now are presented in a Table.  ‘nz-table’ in the nz-content ‘inner-content’ area.
Early column sorting functions added.
Table style tweaks also added.

‘maindb_queries.ts’ file
The * in the ‘Select a.*’ in the ‘getUncategorizedArticles()’ function, replaced by the fields/columns names, so, it includes ‘a.linkurl AS link’ and  ‘a.imageurl AS image’ and thus it is aligned with the corresponding properties of the PostData interface.