feat: introduced a new MarkViewer component opened in a separate Electron window via a new open-new-window IPC flow, updated routing (mark-show root path), added root <router-outlet>, and wired DbArticles to launch it with openWindow(PostData).
‘MarkViewer’ Component (‘mark-viewer.ts’ file)
The new ‘MarkViewer’ Component added. It is loaded in a new Electron window, listens to data received (as a PostData object) in the ngOnInit lifecycle-hook method, extracts the markdown encoded text from the ‘content’ property, parses it in HTML, and finally renders it in its template.

‘app.route.ts’
The path routes updated. All routes for the components that will be rendered into the Layout component are defined as child routes of the Layout component. The new path ‘mark-show’ for the ‘MarkViewer’ component is defined as root path.

‘App’ component (‘app.ts’ file)
The <router-outlet /> is also added in the 'app-root' template (‘app.html’ file) for handling root paths. The already existing <router-outlet /> in the Layout component (inside nz-content of its template), handles only the Layout child routes defined in the ‘app.route.ts’ file.

Electron ‘main.ts’.
The new IPC handler ‘open-new-window’ is defined, for opening a new window, which also accepts the data parameter (a data object).  It instantiates a new Electron window, load the component defined via a ‘hash-prefixed’ route path (‘mark-show’), shows the opened window, and sends the data after the component is loaded.

‘preload.ts’
1 new method: ‘openWindow(data: any)’ added. The ‘openWindow(data)’ triggers Electron to open a new window and passes it the data parameter.

‘electron-api.d.ts’
The ‘ ‘openWindow’ declared in the global electron-api declarations file.

‘DbArticles’ Component
The ‘onMarkdown()’ method update to call the ‘openWindow’ and pass a PostData object.