feat: -globally loaded-  ‘markdown-styles.scss’ file and other updates
‘MarkViewe’r component
MarkViewer component updated and now uses the MarkShow Service for rendering the Markdown content.

‘markdown-styles.scss’ file
We created a separate -globally loaded- ‘markdown-styles.scss’ file, for all Markdown to HTML-rendered cases. It is primarily used in Markdown and MarkViewer components. It is based on the previous ‘.markdown-container’ class used only in the Markdown component. Now updates added for image resizing and other tweaks.

Electron ‘main.ts’ file
The ‘right-click’ context menu with full copy, is also attached to the new window: ‘markViewerWindow’.