refact: The new component ‘ArticlesTable’ created as a component that it is the only one that deals with an ‘nz-table’ for showing Articles from main DB.
The BackEnd service updated and now supports the $articles signal (holding a PostData[] array) that is updated whenever the new ‘setUncategorizedArticlesSignal()’ function is called (from DbArticles component).
The new ‘setUncategorizedArticlesSignal()’ function calls the existing ‘getUncategorizedArticles()’ function to
fetch the PostData[] array with the rows from the ‘articles’ table of the main DB. The $articles signal is obtained in the new ‘ArticlesTable’ component to ‘feed’ the rows of the ‘nz-table’.

The ‘DbArticles’ component refactored accordingly. Any module, variable, template, or style element, related to ‘nz-table’ has been removed. Now in the ‘nz-content’ section (of the ‘nz-layout’ component it uses), it just holds the selector ‘<articles-table />’ of the new ‘ArticlesTable’ component. Upon its initialization (within ngOnInit), it calls the new ‘setUncategorizedArticlesSignal()’ function  of the BackeEnd service that it sets its value to the array of PostData fetched from the ‘articles’ table of the main DB.

The new component ‘ArticlesTable’, upon its instantiation (from within its constructor and the effect() signal-handler), obtains the array of articles (PostData[]) as the value of the BackEnd service $articles signal. Then it uses to feed the row of the ‘nz-table’. Improvements also made in in the MD column, which now shows the ‘nz-icon’ ‘nzType’="file-markdown" with the ‘nzTheme’="outline" if there is no content (or the content is empty) of an article, instead of ‘nzTheme’="twotone".

Other refactoring and updates also took place here and there.