refact/feat: introduced CategoryNodes service and CategoriesTree component to centralize category tree state via signals, refactored DbArticles to use it, and added category selection to SingleUrl with an nz-tree-select.
The new CategoryNodes Service (src/app/shared/services/category-nodes.ts) is created. The ‘$catTreeNodes’ signal is used to keep an array of NzTreeNodeOptions objects. The value of the ‘$catTreeNodes’ signal is set in the ‘getCategoryForestByParentId()’ method that gets an array of CategoryNode objects, and maps them into an array of NzTreeNodeOptions objects (via the mapCategoryNodesToTree() method). Finally, the public method ‘setCategoryTreeNodesSignal()’ can be called from other components to trigger the process of setting the value of the ‘$catTreeNodes’ signal.

The new CategoriesTree Component (src/app/categories-tree/categories-tree.ts) is created. It is only used as an ‘nz-tree’ tree node provider that is populated by an array of NzTreeNodeOptions objects. For this purpose, it monitors for changes and captures the value of the ‘$catTreeNodes’ of the CategoryNodes Service. It does this inside the effect() code block in the CategoriesTree Component constructor.

The DbArticles Component refactored accordingly. The ‘setCategoryTreeNodesSignal()’ of the CategoryNodes Service is called, upon DbArticles components initialization (ngOnInit), to set the value of the ‘$catTreeNodes’ signal. Note that the removed function ‘mapCategoryNodesToTree()’ is actually moved to the CategoryNodes Service, as ‘mapCategoryNodesToTree()’ method.

The SingleUrl component is updated. A new ‘nz-form’ item has been added. This is the ‘nz-tree-select’ form field and it is being filled by category-modes via the ‘$catTreeNodes’ signal of the CategoryNodes Service. It provides a category selection option for the article scraped.