refact/feat: added category–article linking support with new DB queries and IPC handlers, introduced $articles and $selectedCategoryId signals in BackEnd to drive ArticlesTable and CategoriesTree, refactored DbArticles and SingleUrl to handle uncategorized vs category–assigned articles, and wired category selection/assignment across components via signals.
maindb_queries.ts
The new functions: getArticlesByCategoryId(), insertArticleCategory(), deleteArticleCategory(),  deleteArticlesByCategoryId(), and insertArticleCategories(), added.

Electron main.ts
The new 'sqlite:get-articles-by-category-id' and 'sqlite:insert-Article-Categories' IPC channel handlers add for the getArticlesByCategoryId() and the insertArticleCategories() functions respectively.

BackEnd service
The BackEnd service sets and provides 2 signals: the $articles and the $selectedCategoryId. The first provides an array of PostData objects and the second the number of the last selected (clicked) category. The $articles signal is used in the ArticlesTable comp., in which it sets the corresponding signal $articles to feed and form an nz-table. The $selectedCategoryId signal is set in the CategoriesTree comp (in its onTreeNodeClick() method), and then, the backendService.setArticlesByCategoryIdSignal() method is called to update the $articles array by selecting only the articles belonging to the selected category.
The BackEnd $selectedCategoryId signal is used in the DbArticles comp (in its isOnlyUnassignedToggle() method) and if its value is > 0, the backendService.setArticlesByCategoryIdSignal() method is called to update the $articles array by selecting only the articles belonging to the selected category. It is also set to 0 when the DBArticles component is destroyed.
The new methods: setArticlesByCategoryIdSignal(), getArticlesByCategoryId() and insertArticleCategories() also added. The getArticlesByCategoryId() calls the IPC channel handler: 'sqlite:get-articles-by-category-id' and the insertArticleCategories calls the 'sqlite:insert-Article-Categories' IPC channel handler.

CategoryNodes service
The CategoryNodes service just provides the $catTreeNodes signal which is actually, an array of a category ‘forest’ nodes, mapped to an array of NzTreeNodeOptions[] objects. The $catTreeNodes signal is used in the CategoriesTree component, in which it sets the corresponding signal $treeNodes to feed and form an nz-tree array of category nodes.

DbArticles comp.
The DbArticles comp uses both the CategoriesTree and the ArticlesTable comp. It also uses the BackEnd service.
The DbArticles comp. initially calls the backendService.setUncategorizedArticlesSignal() that sets the $articles signal with all unassigned with a category, articles.
It does the same whenever the unassignedOnly Boolean variable is getting true (in isOnlyUnassignedToggle() method). But, if the unassignedOnly value is false (and if the value of the BackEnd $selectedCategoryId signal is > 0, it calls the backendService.setArticlesByCategoryIdSignal() method to update the $articles array by selecting only the articles belonging to the selected category.

CategoriesTree comp
CategoriesTree comp. uses the CategoryNodes service and the BackEnd service.
The CategoriesTree comp. gets its categories’ nodes (for all defined categories), via the CategoryNodes service, as a signal -the $treeNodes signal- in the effect() code block in its constructor, obtained from the corresponding $catTreeNodes of the CategoryNodes service.
The CategoriesTree comp (in its onTreeNodeClick() method), sets the BackEnd $selectedCategoryId signal and then calls the backEndService.setArticlesByCategoryIdSignal() that sets the $articles signal with the articles for a selected category – a category clicked on the categories-tree-nodes

SingleUrl comp
The SingleUrl comp updated. The nz-tree-select obtains its categories nodes using the component’s $categoryNodes signal, which is set by the $catTreeNodes signal of the CategoryNodes service, in the effect() code block, of the SingleUrl comp constructor.
The user selected category and its key value (category id) is captured using a subscription to the form’s valueChanges event for the selectCategory form field (in the ngAfterViewInit() lifecycle hook method). The key (category id) is passed in the component’s property ‘selectedCategoryIds’ and the it is used in the insertScrapedArticlesArrayToDB() and updateScrapedArticleById() methods, as parameter for calling the BackEnd service insertArticleCategories() method.