feat: A new nz-drawer supported by the new ArticleCategoriesSelection component added. It is called from the ArticlesTable host component, allowing to select/unselect the categories assigned to an article.
ArticleCategoriesSelection component
The new ArticleCategoriesSelection component is responsible mainly to provide the nz-tree to nz-drawer, and handle the nz-tree’s main functionalities, e.g., onCheck(), onTreeNodeClick(), onTreeNodeExpandChange(). Moreover, it also manages the canceling and closing the drawer, i.e., cancel() and applyAndClose() via the drawer reference injected in the component and the corresponding buttons. Additionally, it also expands the parent nodes (via the setExpandedKeys() method) for each of the initialKeys provided, which is important for easily spotting them in the nz-tree. Note, that the initialKeys are obtained in the component via the @Input annotation, which captures them from the nzContentParams { initialKeys } of the openCategoryDrawer() method in the ArticlesTable host component.

ArticlesTable component.
A new icon-button: nz-icon - nzType="tags" added in the ArticlesTable component, with the ‘openCategoryDrawer()’ handler function. The ‘initialKeys’ (=the already assigned categories) are fetched by the BackEnd service getCategoryIdsOfAnArticle() method. Correspondingly, the ‘selectedKeys’ (= the final selection of the categories) updates the article_categories table by calling the BackEnd service updateArticleCategories() method.
The openCategoryDrawer() method is triggered by clicking on the new column with the nz-icon ‘tags’. The method is mainly responsible to create/open the nz-drawer and pass it the ArticleCategoriesSelection component , as well as to pass to the the ArticleCategoriesSelection component parameters like the nzContentParams { initialKeys }. The method gets the already assigned categories of an article by the BackEnd service getCategoryIdsOfAnArticle() method. The method also subscribes to the afterClose method of the drawer reference and gets the final selection of checked nodes (selected categories). Then updates the article_categories table accordingly by calling the BackEnd service updateArticleCategories() method.

BackEnd service
The new methods: getCategoryIdsOfAnArticle(), deleteAllArticleCategories(), and updateArticleCategories(), added to the BackEnd service and they are respectively linked to the following IPC channel handlers: 'sqlite:get-category-ids-of-article', 'sqlite:delete-all-article-categories', and ‘sqlite:update-article-categories' in the Electron main.ts file.

Electron main.ts
The Electron main.ts updated with the following IPC channel handlers: 'sqlite:get-category-ids-of-article',
'sqlite:delete-all-article-categories', and ‘sqlite:update-article-categories' which are responsible to call respectively, the functions: getCategoryIdsOfAnArticle(), deleteAllArticleCategories(), and updateArticleCategories(), in the maindb_queries.ts file.

maindb_queries.ts file
Correspondingly, the functions: getCategoryIdsOfAnArticle(), deleteAllArticleCategories(), and updateArticleCategories().

Angular app.config.sys
The Angular app.config.sys updated and now the HttpClient provider uses the withFetch(): provideHttpClient(withFetch()),

Angular main.ts
The Angular main.ts file updated with icon modules: TagsTwoTone and TagsOutline.