upd/feat/refactor: introduced category-aware signals ($selectedCategory, $categoriesFilter) with getCategoryById support, updated IPC, BackEnd, and components (ArticlesTable, CategoriesTree, DbArticles) to use setSelectedCategorySignal, and added table header with refresh support.
maindb_queries.ts
The getCategoryById() function added.

Electron main.ts
The new IPC channel handler ‘’ added for passing a call to the getCategoryById() function in the maindb_queries.ts

BackEnd service
The signal $selectedCategoryId replaced by the new $selectedCategory signal
The new signal: $categoriesFilter also added.
The method getCategoryById() added.
The method setSelectedCategorySignal() added that calls/uses the getCategoryById() method and sets the $selectedCategory signal.
So, the ‘callers’ that used to set directly the value of the $selectedCategoryId signal, now they must call the setSelectedCategorySignal(). They also must use the $selectedCategory().id instead of just the $selectedCategoryId signal

ArticlesTable
A new table header added in the component’s template, providing info about the categories filter (and the category name selected) and the Number of articles fetched.
The new component property named ‘filter’ added and obtain its value by monitoring and capturing the BackEnd $categoriesFilter signal.
The method: updateArticlesTable(), added. It is called after updating the article_categories table (after the selection of categories of an article - by clicking the tags column and using the drawer).

CategoriesTree
Updated to use the setSelectedCategorySignal(). (Previously used to use directly the $selectedCategoryId signal).

DbArticles
Updated to use the setSelectedCategorySignal instead of previously used $selectedCategoryId signal. It also sets the value of the BackEnd $categoriesFilter (inside the onArticlesFilterChange() method).

Angular main.ts
The ReloadOutline nz-icon module added (used in the Refresh button of the new header of the articles table).