feat: a right-click context added -inside the drawer- that provides a pop-up-form for entering a new (sub)category in the categories table, during the category selection in the drawer.
maindb_queries.ts
The new addNewCategory() function added. It adds a new category to the database.  Returns the created category or null if it is failed.

Electron main.ts
The new IPC channel handler ‘sqlite:add-new-category' added for passing a call to the addNewCategory()  function in the maindb_queries.ts

BackEnd service
The addNewCategory() method added that calls the ‘sqlite:add-new-category' IPC channel handler in Electron main.ts.

ArticleCategoriesSelection
The component is heavily updated:
-	It provides a right-click context (NzContextMenuService) -inside the drawer- that adds (inserts) a sub-category in the categories table, during the category selection in the drawer.
-	By right-clicking on the context entry, a modal (NzModalService, NzModalRef) pop-up form is displayed for entering the name of the new subcategory to be added.

The following handler methods added:
-	rememberContext()
-	openAddSubcategoryModal()
-	onRightClick()
-	popUpFormSubmitAdd()
-	onTreeEvent()

We use the [nzTreeTemplate]. It lets us replace the label/content area of every tree node with our own <ng-template>. Here, we use it, just for allowing us to capture the right-click on it for popping-up a context area. The template receives a context with: a) $implicit → an NzTreeNode (runtime node state) and b) origin → your original NzTreeNodeOptions object (the data you passed in). We must use the let-node and let-origin="origin" to access those two values.
The context area (the contextMenu) is created in the ‘onRightClick()’ method. Note that, we must already have injected the NzContextMenuService in our component. Then, we must pass in it, a parameter (e.g.,’ ctxMenu’) that concerns the drop-down area (nz-dropdown-menu) for the context menu item(s) to be displayed.
The right-click event is handled in the ‘openAddSubcategoryModal()’ method. This method, is responsible to pop-up the modal window, and it actually passes in it another template as its content (nzContent). In our case, this template is the ‘addSubcategoryTpl’ template, which is the place to deploy the popUpForm, for entering a new (sub)category.

The key job is done in the popUpFormSubmitAdd() method, after the popUpForm submission. It calls the BackEnd addNewCategory() to create/add a new category in the categories table. Then, it re-sets the $treeNodes signal by calling the setCategoryTreeNodesSignal() method  of the CategoryNodes service, to include the newly added category. After this, it adds the id of the newly created category to the $initialKeys, and re-sets the $initialKeyStrings, $checkedKeys signals. Finally, it calls the component’s method  setExpandedKeys(), and ‘pre-selects’ the added category, so it will be ‘visible’ for selection.

Note:
The $initialKeyStrings computed signal replaced by a ‘normal’ signal, since it wasn’t been updated automatically by the @Inputed initialKey.